Updates
I've done the search function for a practical user not an admin.  IT
works fine.  I also modified the databases as well since we might not
want the username to be the primary key I've set up Auto-incremented
values for each one to be the primary key.  I've also set the username
to Unique so we can refernence it.  The highest_bidder I am having
trouble with as it will not let it be null at the beggining as we don't
know who the highest bidder is.  I am now going to try to get actual
numbers inserted into the database for min_bid and decimal purposes.